deal with the rename of "our" debian directory fix typo in init.d script (Matthijs) fix compilation error with g++ 4.0 on AMD64 (Matthijs)
git-svn-id: svn://svn.powerdns.com/pdns/trunk/pdns@341 d19b8d6e-7fed-0310-83ef-9ca221ded41b

+++ pdns/tcpreceiver.cc
@@ -129,7 +129,8 @@ void TCPNameserver::getQuestion(int fd, char *mesg, int pktlen, const struct soc
 
 void *TCPNameserver::doConnection(void *data)
 {
-  int fd=(int)data; // gotta love C (generates a harmless warning on opteron)
+  // Fix gcc-4.0 error (on AMD64)
+  int fd=(int)(long)data; // gotta love C (generates a harmless warning on opteron)
   pthread_detach(pthread_self());
 
   try {
